Move cloud commands to root
Moves all cloud commands to root level as they should be used for
interacting with Upbound, not just Upbound Cloud.
