+++ cmd/up/controlplane/delete.go
