+++ cmd/up/controlplane/kubeconfig/get.go
