+++ cmd/up/controlplane/token/delete.go
