+++ cmd/up/controlplane/token/token.go
